Move providers logic to dedicated service
