+++ plugins/woocommerce/src/Internal/Admin/Settings/PaymentsController.php
@@ -186,7 +186,14 @@ class PaymentsController {
 		foreach ( $providers as $provider ) {
 			// Handle payment gateways and offline payment methods that need setup.
 			if (
-				in_array( $provider['_type'], array( Payments::PROVIDER_TYPE_GATEWAY, Payments::PROVIDER_TYPE_OFFLINE_PM ), true ) &&
+				in_array(
+					$provider['_type'],
+					array(
+						PaymentProviders::TYPE_GATEWAY,
+						PaymentProviders::TYPE_OFFLINE_PM,
+					),
+					true
+				) &&
 				! empty( $provider['state']['needs_setup'] )
 			) {
 				return true;
