+++ plugins/woocommerce/src/Internal/Admin/Settings/PaymentsRestController.php
@@ -52,6 +52,28 @@ class PaymentsRestController extends RestApiControllerBase {
 	public function register_routes( bool $override = false ) {
 		register_rest_route(
 			$this->route_namespace,
+			'/' . $this->rest_base . '/country',
+			array(
+				array(
+					'methods'             => \WP_REST_Server::EDITABLE,
+					'callback'            => fn( $request ) => $this->run( $request, 'set_country' ),
+					'validation_callback' => 'rest_validate_request_arg',
+					'permission_callback' => fn( $request ) => $this->check_permissions( $request ),
+					'args'                => array(
+						'location' => array(
+							'description'       => esc_html__( 'The ISO3166 alpha-2 country code to save for the current user.', 'woocommerce' ),
+							'type'              => 'string',
+							'pattern'           => '[a-zA-Z]{2}', // Two alpha characters.
+							'required'          => true,
+							'validate_callback' => fn( $value, $request ) => $this->check_location_arg( $value, $request ),
+						),
+					),
+				),
+			),
+			$override
+		);
+		register_rest_route(
+			$this->route_namespace,
 			'/' . $this->rest_base . '/providers',
 			array(
 				array(
@@ -70,28 +92,6 @@ class PaymentsRestController extends RestApiControllerBase {
 					),
 				),
 				'schema' => fn() => $this->get_schema_for_get_payment_providers(),
-			),
-			$override
-		);
-		register_rest_route(
-			$this->route_namespace,
-			'/' . $this->rest_base . '/country',
-			array(
-				array(
-					'methods'             => \WP_REST_Server::EDITABLE,
-					'callback'            => fn( $request ) => $this->run( $request, 'set_country' ),
-					'validation_callback' => 'rest_validate_request_arg',
-					'permission_callback' => fn( $request ) => $this->check_permissions( $request ),
-					'args'                => array(
-						'location' => array(
-							'description'       => esc_html__( 'The ISO3166 alpha-2 country code to save for the current user.', 'woocommerce' ),
-							'type'              => 'string',
-							'pattern'           => '[a-zA-Z]{2}', // Two alpha characters.
-							'required'          => true,
-							'validate_callback' => fn( $value, $request ) => $this->check_location_arg( $value, $request ),
-						),
-					),
-				),
 			),
 			$override
 		);
@@ -190,13 +190,13 @@ class PaymentsRestController extends RestApiControllerBase {
 		$offline_payment_providers = array_values(
 			array_filter(
 				$providers,
-				fn( $provider ) => Payments::PROVIDER_TYPE_OFFLINE_PM === $provider['_type']
+				fn( $provider ) => PaymentProviders::TYPE_OFFLINE_PM === $provider['_type']
 			)
 		);
 		$providers                 = array_values(
 			array_filter(
 				$providers,
-				fn( $provider ) => Payments::PROVIDER_TYPE_OFFLINE_PM !== $provider['_type']
+				fn( $provider ) => PaymentProviders::TYPE_OFFLINE_PM !== $provider['_type']
 			)
 		);
 
@@ -474,7 +474,7 @@ class PaymentsRestController extends RestApiControllerBase {
 
 			// If this is a suggestion, add a link to hide it.
 			if ( ! empty( $provider['_type'] ) &&
-				Payments::PROVIDER_TYPE_SUGGESTION === $provider['_type'] &&
+				PaymentProviders::TYPE_SUGGESTION === $provider['_type'] &&
 				! empty( $provider['_suggestion_id'] )
 				) {
 				$providers[ $key ]['_links']['hide'] = array(
@@ -483,9 +483,11 @@ class PaymentsRestController extends RestApiControllerBase {
 			}
 
 			// If we have an incentive, add a link to dismiss it.
-			if ( ! empty( $provider['_incentive'] ) &&
-				! empty( $provider['_suggestion_id'] )
-				) {
+			if ( ! empty( $provider['_incentive'] ) && ! empty( $provider['_suggestion_id'] ) ) {
+				if ( empty( $provider['_incentive']['_links'] ) ) {
+					$providers[ $key ]['_incentive']['_links'] = array();
+				}
+
 				$providers[ $key ]['_incentive']['_links']['dismiss'] = array(
 					'href' => rest_url( sprintf( '/%s/%s/suggestion/%s/incentive/%s/dismiss', $this->route_namespace, $this->rest_base, $provider['_suggestion_id'], $provider['_incentive']['id'] ) ),
 				);
@@ -634,7 +636,7 @@ class PaymentsRestController extends RestApiControllerBase {
 					'properties'  => array(
 						'_type'  => array(
 							'type'        => 'string',
-							'enum'        => array( Payments::EXTENSION_TYPE_WPORG ),
+							'enum'        => array( PaymentProviders::EXTENSION_TYPE_WPORG ),
 							'description' => esc_html__( 'The type of the plugin.', 'woocommerce' ),
 							'context'     => array( 'view', 'edit' ),
 							'readonly'    => true,
@@ -653,7 +655,11 @@ class PaymentsRestController extends RestApiControllerBase {
 						),
 						'status' => array(
 							'type'        => 'string',
-							'enum'        => array( Payments::EXTENSION_NOT_INSTALLED, Payments::EXTENSION_INSTALLED, Payments::EXTENSION_ACTIVE ),
+							'enum'        => array(
+								PaymentProviders::EXTENSION_NOT_INSTALLED,
+								PaymentProviders::EXTENSION_INSTALLED,
+								PaymentProviders::EXTENSION_ACTIVE,
+							),
 							'description' => esc_html__( 'The status of the plugin.', 'woocommerce' ),
 							'context'     => array( 'view', 'edit' ),
 							'readonly'    => true,
@@ -727,11 +733,26 @@ class PaymentsRestController extends RestApiControllerBase {
 					'type'        => 'object',
 					'description' => esc_html__( 'The management details of the provider.', 'woocommerce' ),
 					'properties'  => array(
-						'settings_url' => array(
-							'type'        => 'string',
-							'description' => esc_html__( 'The URL to the settings page for the payment gateway.', 'woocommerce' ),
-							'context'     => array( 'view', 'edit' ),
-							'readonly'    => true,
+						'_links' => array(
+							'type'       => 'object',
+							'context'    => array( 'view', 'edit' ),
+							'readonly'   => true,
+							'properties' => array(
+								'settings' => array(
+									'type'        => 'object',
+									'description' => esc_html__( 'The link to the settings page for the payment gateway.', 'woocommerce' ),
+									'context'     => array( 'view', 'edit' ),
+									'readonly'    => true,
+									'properties'  => array(
+										'href' => array(
+											'type'        => 'string',
+											'description' => esc_html__( 'The URL to the settings page for the payment gateway.', 'woocommerce' ),
+											'context'     => array( 'view', 'edit' ),
+											'readonly'    => true,
+										),
+									),
+								),
+							),
 						),
 					),
 				),
@@ -739,6 +760,27 @@ class PaymentsRestController extends RestApiControllerBase {
 					'type'        => 'object',
 					'description' => esc_html__( 'Onboarding-related details for the provider.', 'woocommerce' ),
 					'properties'  => array(
+						'_links'                      => array(
+							'type'       => 'object',
+							'context'    => array( 'view', 'edit' ),
+							'readonly'   => true,
+							'properties' => array(
+								'onboard' => array(
+									'type'        => 'object',
+									'description' => esc_html__( 'The start/continue onboarding link for the payment gateway.', 'woocommerce' ),
+									'context'     => array( 'view', 'edit' ),
+									'readonly'    => true,
+									'properties'  => array(
+										'href' => array(
+											'type'        => 'string',
+											'description' => esc_html__( 'The URL to start/continue onboarding for the payment gateway.', 'woocommerce' ),
+											'context'     => array( 'view', 'edit' ),
+											'readonly'    => true,
+										),
+									),
+								),
+							),
+						),
 						'recommended_payment_methods' => array(
 							'type'        => 'array',
 							'description' => esc_html__( 'The list of recommended payment methods details for the payment gateway.', 'woocommerce' ),
@@ -870,20 +912,36 @@ class PaymentsRestController extends RestApiControllerBase {
 							),
 						),
 						'_links'            => array(
-							'type'     => 'array',
-							'context'  => array( 'view', 'edit' ),
-							'readonly' => true,
-							'dismiss'  => array(
-								'type'        => 'object',
-								'description' => esc_html__( 'The link to dismiss the incentive.', 'woocommerce' ),
-								'context'     => array( 'view', 'edit' ),
-								'readonly'    => true,
-								'properties'  => array(
-									'href' => array(
-										'type'        => 'string',
-										'description' => esc_html__( 'The URL to dismiss the incentive.', 'woocommerce' ),
-										'context'     => array( 'view', 'edit' ),
-										'readonly'    => true,
+							'type'       => 'object',
+							'context'    => array( 'view', 'edit' ),
+							'readonly'   => true,
+							'properties' => array(
+								'dismiss' => array(
+									'type'        => 'object',
+									'description' => esc_html__( 'The link to dismiss the incentive.', 'woocommerce' ),
+									'context'     => array( 'view', 'edit' ),
+									'readonly'    => true,
+									'properties'  => array(
+										'href' => array(
+											'type'        => 'string',
+											'description' => esc_html__( 'The URL to dismiss the incentive.', 'woocommerce' ),
+											'context'     => array( 'view', 'edit' ),
+											'readonly'    => true,
+										),
+									),
+								),
+								'onboard' => array(
+									'type'        => 'object',
+									'description' => esc_html__( 'The start/continue onboarding link for the payment gateway.', 'woocommerce' ),
+									'context'     => array( 'view', 'edit' ),
+									'readonly'    => true,
+									'properties'  => array(
+										'href' => array(
+											'type'        => 'string',
+											'description' => esc_html__( 'The URL to start/continue onboarding for the payment gateway.', 'woocommerce' ),
+											'context'     => array( 'view', 'edit' ),
+											'readonly'    => true,
+										),
 									),
 								),
 							),
@@ -891,20 +949,22 @@ class PaymentsRestController extends RestApiControllerBase {
 					),
 				),
 				'_links'            => array(
-					'type'     => 'array',
-					'context'  => array( 'view', 'edit' ),
-					'readonly' => true,
-					'hide'     => array(
-						'type'        => 'object',
-						'description' => esc_html__( 'The link to hide the suggestion.', 'woocommerce' ),
-						'context'     => array( 'view', 'edit' ),
-						'readonly'    => true,
-						'properties'  => array(
-							'href' => array(
-								'type'        => 'string',
-								'description' => esc_html__( 'The URL to hide the suggestion.', 'woocommerce' ),
-								'context'     => array( 'view', 'edit' ),
-								'readonly'    => true,
+					'type'       => 'object',
+					'context'    => array( 'view', 'edit' ),
+					'readonly'   => true,
+					'properties' => array(
+						'hide' => array(
+							'type'        => 'object',
+							'description' => esc_html__( 'The link to hide the suggestion.', 'woocommerce' ),
+							'context'     => array( 'view', 'edit' ),
+							'readonly'    => true,
+							'properties'  => array(
+								'href' => array(
+									'type'        => 'string',
+									'description' => esc_html__( 'The URL to hide the suggestion.', 'woocommerce' ),
+									'context'     => array( 'view', 'edit' ),
+									'readonly'    => true,
+								),
 							),
 						),
 					),
@@ -967,7 +1027,7 @@ class PaymentsRestController extends RestApiControllerBase {
 					'properties' => array(
 						'_type'  => array(
 							'type'        => 'string',
-							'enum'        => array( Payments::EXTENSION_TYPE_WPORG ),
+							'enum'        => array( PaymentProviders::EXTENSION_TYPE_WPORG ),
 							'description' => esc_html__( 'The type of the plugin.', 'woocommerce' ),
 							'context'     => array( 'view', 'edit' ),
 							'readonly'    => true,
@@ -980,7 +1040,11 @@ class PaymentsRestController extends RestApiControllerBase {
 						),
 						'status' => array(
 							'type'        => 'string',
-							'enum'        => array( Payments::EXTENSION_NOT_INSTALLED, Payments::EXTENSION_INSTALLED, Payments::EXTENSION_ACTIVE ),
+							'enum'        => array(
+								PaymentProviders::EXTENSION_NOT_INSTALLED,
+								PaymentProviders::EXTENSION_INSTALLED,
+								PaymentProviders::EXTENSION_ACTIVE,
+							),
 							'description' => esc_html__( 'The status of the plugin.', 'woocommerce' ),
 							'context'     => array( 'view', 'edit' ),
 							'readonly'    => true,
